+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * `plom-server launch --mastertoken aabbccdd...` replaces the old way (without the keyword argument).
+* QR creation now uses `segno` instead of `pyqrcode`.
 
 ### Fixed
 * `plom-finish reassemble` not longer needs direct file access to the server (except when using the `--ided-only` option).